requirements.txt
allure-pytest==2.13.2
Faker==20.1.0
pytest==7.4.3
requests==2.31.0
selenium==4.15.2
